Analysis: Implement `@herb-tools/analysis` package
The pieces that answer "what templates does this project have, and which ones
render which" were split across two packages that had no business owning them.
Core carried the index data structures, so a browser-only consumer of the parser
dragged the whole template graph along with it. The linter carried the builders
that fill those indexes, so the language server depended on the linter in order
to index a project it was not linting.

`@herb-tools/analysis` now owns both halves. The root entry is browser-safe and
holds the data structures and the pure collectors, while `@herb-tools/analysis/node`
holds the parts that read from disk, since building an index means globbing for
templates.

Two helpers come along because the call site collector is their only caller
outside a handful of rules: the readers that decide which argument of a render
call names a template, and the ancestor attribute collector. The linter now
imports them back from analysis.

A third one turned out to be unnecessary. `isOutputRender` did what core's
`isERBOutputNode` already does, so it is gone and its four callers use the one
in core, which narrows the node type rather than only answering yes or no.

Action View is the only framework whose conventions are encoded here today, but
nothing about the package is bound to it. Herb already recognises Hanami and
Sinatra, and their templates raise the same questions.

This is a move. No behavior changes, no serialized format, no dependency
tracking. Those want their own pull requests, and the second one has to agree
with the Ruby implementation on `origin/dependency-tracking` rather than being
invented here.
