+++ .github/labeler.yml
@@ -172,6 +172,11 @@ language-service:
     - any-glob-to-any-file:
       - '**/javascript/packages/language-service/**/*'
 
+analysis:
+  - changed-files:
+    - any-glob-to-any-file:
+      - '**/javascript/packages/analysis/**/*'
+
 printer:
   - changed-files:
     - any-glob-to-any-file:
