+++ javascript/packages/analysis/src/partial-resolution.ts
