+++ javascript/packages/language-server/package.json
@@ -45,6 +45,7 @@
     "dist/"
   ],
   "dependencies": {
+    "@herb-tools/analysis": "0.10.3",
     "@herb-tools/config": "0.10.3",
     "@herb-tools/core": "0.10.3",
     "@herb-tools/formatter": "0.10.3",
