+++ javascript/packages/language-server/project.json
@@ -10,6 +10,7 @@
         "script": "build"
       },
       "dependsOn": [
+        "@herb-tools/analysis:build",
         "@herb-tools/config:build",
         "@herb-tools/core:build",
         "@herb-tools/formatter:build",
