+++ javascript/packages/linter/package.json
@@ -44,19 +44,10 @@
       "import": "./dist/loader.js",
       "require": "./dist/loader.cjs",
       "default": "./dist/loader.js"
-    },
-    "./partial-index-builder": {
-      "types": "./dist/types/partial-index-builder.d.ts",
-      "import": "./dist/partial-index-builder.js",
-      "default": "./dist/partial-index-builder.js"
-    },
-    "./partial-caller-builder": {
-      "types": "./dist/types/partial-caller-builder.d.ts",
-      "import": "./dist/partial-caller-builder.js",
-      "default": "./dist/partial-caller-builder.js"
     }
   },
   "dependencies": {
+    "@herb-tools/analysis": "0.10.3",
     "@herb-tools/config": "0.10.3",
     "@herb-tools/core": "0.10.3",
     "@herb-tools/highlighter": "0.10.3",
