+++ javascript/packages/linter/src/cli/file-processor.ts
@@ -12,17 +12,18 @@ import { availableParallelism } from "node:os"
 import { colorize } from "@herb-tools/highlighter"
 import { deserializeDiagnostic, didyoumean } from "@herb-tools/core"
 import { fixabilityFor } from "../fixability.js"
-import { buildPartialIndex, refreshPartialAfterFix } from "../partial-index-builder.js"
-import { buildPartialCallerIndex } from "../partial-caller-builder.js"
+import { buildPartialIndex, refreshPartialAfterFix } from "@herb-tools/analysis/node"
+import { buildPartialCallerIndex } from "@herb-tools/analysis/node"
 
-import type { AncestorChain, Diagnostic } from "@herb-tools/core"
+import type { Diagnostic } from "@herb-tools/core"
+import type { AncestorChain } from "@herb-tools/analysis"
 import type { FormatOption } from "./argument-parser.js"
 import type { HerbConfigOptions } from "@herb-tools/config"
 import type { WorkerInput, WorkerResult } from "./lint-worker.js"
 import type { Fixability } from "../fixability.js"
 import type { VersionSkippedRule } from "../linter.js"
 import type { LintOffense, RuleClass } from "../types.js"
-import type { PartialCallerIndex, PartialIndex } from "@herb-tools/core"
+import type { PartialCallerIndex, PartialIndex } from "@herb-tools/analysis"
 
 const AUTOMATIC_FIX_DIFF_LIMIT = 20
