+++ javascript/packages/linter/test/rules/html-body-only-elements.test.ts
@@ -1,7 +1,7 @@
 import dedent from "dedent"
 
 import { describe, test } from "vitest"
-import { PartialCallerIndex } from "@herb-tools/core"
+import { PartialCallerIndex } from "@herb-tools/analysis"
 import { createLinterTest } from "../helpers/linter-test-helper.js"
 import { HTMLBodyOnlyElementsRule } from "../../src/rules/html-body-only-elements.js"
 import { LAYOUT, renderedFrom, renderedFromNowhere } from "../helpers/partial-caller-context.js"
